# swagger.yaml
openapi: 3.0.3
info:
  title: Ticker Pricing API
  description: API для получения цен и графиков по тикерам 
  version: 1.0.0
tags:
  - name: Charts
    description: Эндпоинты для построения графиков
  - name: API
    description: Основные эндпоинты для работы с данными
paths:
  /linear:
    get:
      tags:
        - Charts
      summary: Get All For Charts
      description: Возвращает данные для построения линейного графика по тикеру.
      operationId: get_all_for_charts
      parameters:
        - name: ticker
          in: query
          required: true
          description: Ticker symbol
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 200
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LinearPriceChart"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"

  /linear/get_chart_history:
    get:
      tags:
        - Charts
      summary: Get Chart History
      description: Возвращает исторические данные для графика.
      operationId: get_chart_history
      parameters:
        - name: ticker
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 200
        - name: before_timestamp
          in: query
          required: false
          description: Unix timestamp самой старой точки на клиенте
          schema:
            type: integer
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LinearPriceChart"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"

  /all:
    get:
      tags:
        - API
      summary: Get All
      description: Возвращает все записи цен для конкретного тикера с пагинацией.
      operationId: get_all
      parameters:
        - name: ticker
          in: query
          required: true
          description: Ticker symbol (e.g., btc_usd or eth_usd)
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 5
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceResponse"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"

  /latest:
    get:
      tags:
        - API
      summary: Get Latest
      description: Возвращает самую последнюю запись цены для указанного тикера.
      operationId: get_latest
      parameters:
        - name: ticker
          in: query
          required: true
          description: Ticker symbol (e.g., btc_usd or eth_usd)
          schema:
            type: string
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceResponse"
        "404":
          description: No data found for this ticker
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"

  /by_date:
    get:
      tags:
        - API
      summary: Get By Date
      description: Возвращает данные по тикеру в заданном временном диапазоне.
      operationId: get_by_date
      parameters:
        - name: ticker
          in: query
          required: true
          description: Ticker symbol (e.g., btc_usd or eth_usd)
          schema:
            type: string
        - name: start
          in: query
          required: true
          description: Start UNIX timestamp
          schema:
            type: integer
        - name: end
          in: query
          required: false
          description: End UNIX timestamp (optional)
          schema:
            type: integer
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 5
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceResponse"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"

  /tickers_list:
    get:
      tags:
        - API
      summary: Get All Tickers
      description: Возвращает список всех доступных тикеров.
      operationId: get_all_tickers
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TickerName"

components:
  schemas:
    LinearPriceChart:
      type: object
      required:
        - time
        - value
      properties:
        time:
          type: integer
          description: UNIX timestamp точки
        value:
          type: number
          description: Значение цены
          
    PriceResponse:
      type: object
      required:
        - ticker
        - price
        - timestamp
      properties:
        id:
          type: integer
        ticker:
          type: string
          description: Ticker symbol
        price:
          type: number
          description: Значение цены
        timestamp:
          type: integer
          description: UNIX timestamp записи
          
    TickerName:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Имя тикера (например, btc_usd)

    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"
            
    ValidationError:
      type: object
      required:
        - loc
        - msg
        - type
      properties:
        loc:
          type: array
          items:
            type: string
        msg:
          type: string
        type:
          type: string